requirements.txt
Django>=1.5,<1.6
django-tinymce>=1.5,<2.0
South>=0.8,<1.0
MySQL-python>=1.2,<2.0
Pillow>=2.2,<3.0
django-admin-tools>=0.5.1,<1.0.0
feedparser>=5.1.3,<6.0.0
sorl-thumbnail>=11.12,<12.0
python-instagram>=0.8.0,<1.0.0
